Separate hiding behavior and showing snackbar from above
